Using test runner.
